% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_db.R
\name{double_cases}
\alias{double_cases}
\title{Double the cases in a dataset}
\usage{
double_cases(data, y_name = "y")
}
\arguments{
\item{data}{A dataset.}

\item{y_name}{Name or index of the outcome column in \code{data}.}
}
\value{
Returns a dataset where each case in the input dataset is doubled. An
  ID column (\code{.row_id_db}) is created such that the two rows created
  from the same case has the same ID (the name of this new id column begins
  with \code{.} so as to minimise the risk of overwriting an existing column
  in the input data).
}
\description{
Double the cases in a dataset
}
