% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_db.R
\name{logit_db}
\alias{logit_db}
\title{Estimate the relative risk from a binary outcome by using the
doubling-of-cases approach}
\usage{
logit_db(formula, weight_name = NULL, dat)
}
\arguments{
\item{formula}{Formula.}

\item{weight_name}{Name or column index in \code{dat} corresponding to the
sampling weights, if \code{dat} is a case-control sample. Default is
\code{NULL}, corresponding to scenarios where the full cohort (or
cross-sectional data) is available.}

\item{dat}{A dataset (without doubling the cases).}
}
\description{
Estimate the relative risk from a binary outcome by using the
doubling-of-cases approach
}
\details{
This function doubles the cases in \code{dat} by using the
  \code{\link{double_cases}} function, and then applies a logistic regression
  to this modified dataset to estimate the relative risk. The robust SE of
  the estimated log relative risk is described in Shouten et al (1993).

  If \code{dat} is a (matched) case-control sample, the logistic regression
  need to be weighted by the sampling weights corresponding to each subject
  in \code{dat}, where the weight is 1 for each case (if all cases are
  selected), and the weight for each control is the number of eligible
  controls divided by the number of controls sampled (counted within each
  sampling stratum if \code{dat} is a matched sample). The robust SE in such
  analyses is modified from Shouten et al (1993) by incorporating the
  sampling weights in H2.
}
\references{
\itemize{
  \item{Schouten, E.G., Dekker, J.M., Kok, F.J., Cessie, S.L., Van
  Houwelingen, H.C., Pool, J. and Vandnbroucke, J.P. (1993). Risk ratio and
  rate ratio estimation in case‐cohort designs: Hypertension and
  cardiovascular mortality. Statist. Med., 12: 1733-1745.
  doi:10.1002/sim.4780121808}
  \item{Knol, M. J., Le Cessie, S., Algra, A., Vandenbroucke, J. P., &
  Groenwold, R. H. (2012). Overestimation of risk ratios by odds ratios in
  trials and cohort studies: alternatives to logistic regression. CMAJ :
  Canadian Medical Association journal, 184(8), 895–899.
  https://doi.org/10.1503/cmaj.101715}
}
}
