% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_db.R
\name{vcov.db}
\alias{vcov.db}
\alias{logLik.db}
\alias{AIC.db}
\alias{summary.db}
\title{Summarise the fitted model for the doubling of cases approach}
\usage{
\method{vcov}{db}(object, ...)

\method{logLik}{db}(object, ...)

\method{AIC}{db}(object, ..., k = 2)

\method{summary}{db}(object, ...)
}
\arguments{
\item{object}{The model fitted using \code{\link{logit_db}}.}

\item{...}{Additional input.}
}
\description{
Summarise the fitted model for the doubling of cases approach
}
\details{
See usage in \code{\link{logit_db}}.
}
\section{Functions}{
\itemize{
\item \code{vcov.db}: Extract corrected covariance matrix

\item \code{logLik.db}: Extract log-likelihood with respect to the log-binomial model of interest

\item \code{AIC.db}: Extract AIC
}}

