% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_db.R
\name{vcov.db}
\alias{vcov.db}
\alias{logLik.db}
\alias{AIC.db}
\alias{confint.db}
\alias{summary.db}
\title{Summarise the fitted model for the doubling of cases approach}
\usage{
\method{vcov}{db}(object, ...)

\method{logLik}{db}(object, ...)

\method{AIC}{db}(object, ..., k = 2)

\method{confint}{db}(object, parm, exp = FALSE, level = 0.95, ...)

\method{summary}{db}(object, ...)
}
\arguments{
\item{object}{The model fitted using \code{\link{logit_db}}.}

\item{...}{Additional input.}

\item{parm}{A specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{exp}{Whether to generate confidence interval for the exponentiated
coefficients (i.e., the estimated risk ratio, if \code{TRUE}, the default),
or for the estimated coefficients (if \code{FALSE}).}

\item{level}{The confidence level required. Default is 0.95 (i.e., 95\%).}
}
\description{
Summarise the fitted model for the doubling of cases approach
}
\details{
See usage in \code{\link{logit_db}}.
}
\section{Functions}{
\itemize{
\item \code{vcov.db}: Extract corrected covariance matrix

\item \code{logLik.db}: Extract log-likelihood with respect to the log-binomial model of interest

\item \code{AIC.db}: Extract AIC

\item \code{confint.db}: Confidence intervals for estimated coefficients
}}

